refactor: consolidate the code between Code and OSS suggestion webview [ROAD-284] (#61)
* refactor: consolidate webview between OSS and Code

* fix: icon and devDependency
