+++ media/views/snykCode/suggestion/suggestion.css
@@ -1,0 +1,97 @@
+html { height: 100%; font-size: 62.5% }
+body { height: 100%; padding: 0; font-size: 1.4rem; line-height: 1.5;  }
+body * { box-sizing: border-box }
+
+.button { vertical-align: middle; padding: .9rem 1.5rem 1rem; border: 1px #7754DB solid; border-radius: 30px; line-height: 1; background: none; color: #7754DB; text-align: center; font-weight: 700; font-family: inherit; transition: all 0.25s; cursor: pointer; }
+.button.disabled { border: 1px #676C6F solid; color: #676C6F; }
+.button:hover { background-color: #7754DB; color: #fff; }
+.button.disabled:hover { background-color: #676C6F; }
+
+.font-light { opacity: .75 }
+.font-blue { color: #7754DB; }
+.font-red { color: #FC3838; }
+
+.row { display: flex; flex-direction: row; width: 100%; }
+.row.between { justify-content: space-between; }
+.row.center { justify-content: center; }
+
+.vscode-light { background-color: rgba(0,0,0,.05); border-right: 1px solid rgba(0,0,0,.05); }
+.vscode-dark { background-color: rgba(255,255,255,.075); border-right: 1px solid rgba(255,255,255,.075); }
+.vscode-light .delimiter-top { border-top: 1px solid rgba(0,0,0,.05); }
+.vscode-dark .delimiter-top { border-top: 1px solid rgba(255,255,255,.075); }
+
+.icon { display: inline-flex; vertical-align: middle; width: 18px; height: 18px; }
+
+section { padding: 20px }
+.suggestion { position: relative; display: flex; flex-direction: column; width: 100%; height: 100%; }
+#suggestion-info { padding-left: 12rem }
+.suggestion-text { float:left; margin-bottom: 2rem; font-size:1.6rem; line-height: 1.6; }
+.suggestion-text.critical .mark-position { color: #CE5019 }
+.suggestion-text.warning .mark-position { color: #d68000 }
+.suggestion-text.info .mark-position { color: #88879e }
+.suggestion-text.critical .mark-message:hover { color: #CE5019 }
+.suggestion-text.warning .mark-message:hover { color: #d68000 }
+.suggestion-text.info .mark-message:hover { color: #88879e }
+.mark-message { font-weight: bold; }
+
+#severity { display:flex; flex-direction: column; flex-grow: 0; float:left; width:80px; margin:1rem 0 0 -10rem; text-align: center }
+#severity .icon { width: 32px; height: 32px; margin: 0 auto 10px;  }
+#severity-text {  }
+
+.vscode-dark .light-only { display: none; }
+.vscode-light .dark-only { display: none; }
+.hidden { display: none; }
+.clickable:hover { cursor: pointer; }
+
+.chip { opacity:.75; display: inline-flex; padding: 2px 8px; border-radius: 15px; margin: 3px; background-color: rgba(150, 150, 150, 0.15); text-transform: capitalize; }
+
+#info-top { margin-bottom: 8px; }
+#example-top { margin-bottom: 16px; }
+#explanations-group { padding-top:16px }
+
+.arrow { display: inline-flex; vertical-align: middle; cursor: pointer; }
+.arrow.enabled { fill: #7754DB; color: #7754DB; }
+.arrow:hover { fill: #7754DB; color: #7754DB; }
+.arrow.left { transform: rotate(90deg); }
+.arrow.right { transform: rotate(-90deg); }
+.arrow.down { transform: rotate(-180deg); }
+
+#example { width: 100%; border: 1px solid; border-radius: 3px; line-height: 1.5; background-color: #fff; font-weight: 600 }
+.vscode-light #example { border-color: rgba(0,0,0,.15) }
+.example-line.removed { background-color: #ffeef0; }
+.example-line.removed>code {font-weight:600 }
+.example-line.added { background-color: #e6ffed; }
+.example-line.added>code { font-weight:600 }
+.example-line>code { padding-left: 30px; white-space: pre-wrap; color: #231F20; font-weight:400 }
+#explanations-top { margin-top: 16px; margin-bottom: 8px; }
+
+.vscode-dark #example { border-color: rgba(255,255,255,.075); background-color: rgba(0,0,0,.15)  }
+.vscode-dark .removed { background-color:rgba(201,60,55,0.2); color:#fff }
+.vscode-dark .added { background-color:rgba(70,149,74,0.2); color:#fff }
+.vscode-dark .example-line>code { color: #ccc; font-weight:400 }
+.vscode-dark .added>code { color: #fff; }
+.vscode-dark .removed>code { color: #fff; }
+
+
+#ignore-top { width: 100%; margin-bottom: 8px; text-align: center; }
+.ignore-actions { justify-content: center }
+.ignore-actions .button { margin: 0 1rem 2rem }
+
+#feedback-close { padding: 15px; border-radius: 5px; border: 1px solid; }
+.vscode-light #feedback-close { border-color: rgba(0,0,0,.15); }
+.vscode-dark #feedback-close { border-color: rgba(255,255,255,.075); }
+#feedback-close:hover { background-color: rgba(80, 200, 239, 0.1); }
+#feedback-open { width: 100%; }
+.vscode-light #feedback-textarea { border-color: rgba(0,0,0,.15) }
+.vscode-dark #feedback-textarea { border-color: rgba(255,255,255,.075);  }
+#feedback-textarea { width: 100%; padding:5px; margin-top:1rem; line-height: 1.5 }
+#feedback-sent-message { width: 100%; }
+
+.suggestion-links { display:flex; width: 100%; line-height: 1 }
+#lead-url { margin-left: auto }
+
+.feedback-section { margin-top: auto }
+.false-positive { float:right; }
+
+.feedback-actions { justify-content: center }
+.feedback-actions .button { margin: 2rem 1rem }
