+++ package.json
@@ -114,7 +114,7 @@
         {
           "id": "snyk",
           "title": "Snyk",
-          "icon": "images/snyk_extension_icon_new.svg"
+          "icon": "media/images/snyk_extension_icon_new.svg"
         }
       ]
     },
@@ -293,8 +293,7 @@
     "sinon": "^11.1.1",
     "typescript": "^4.0.2",
     "vscode-test": "^1.4.0",
-    "yalc": "^1.0.0-pre.44",
-    "rxjs": "^7.3.0"
+    "yalc": "^1.0.0-pre.44"
   },
   "dependencies": {
     "@itly/plugin-amplitude-node": "^2.5.0",
@@ -307,6 +306,7 @@
     "lodash": "^4.17.21",
     "open": "^7.4.2",
     "uuid": "^8.3.2",
+    "rxjs": "^7.3.0",
     "marked": "^3.0.2"
   }
 }
