+++ src/snyk/snykCode/views/interfaces.ts
@@ -1,7 +1,7 @@
 import { IExtension } from '../../base/modules/interfaces';
 import * as vscode from 'vscode';
 
-export interface ISuggestionProvider {
+export interface ICodeSuggestionWebviewProvider {
   activate(extension: IExtension): void;
   show(suggestionId: string, uri: vscode.Uri, position: vscode.Range): void;
   checkCurrentSuggestion(): void;
