+++ src/snyk/snykOss/views/ossVulnerabilityTreeProvider.ts
@@ -18,7 +18,7 @@ export type OssIssueCommandArg = OssVulnerability & {
   overviewHtml: string;
 };
 
-export class OssVulnerabilityProvider extends AnalysisTreeNodeProvder {
+export class OssVulnerabilityTreeProvider extends AnalysisTreeNodeProvder {
   constructor(
     protected readonly viewManagerService: IViewManagerService,
     protected readonly contextService: IContextService,
@@ -66,9 +66,9 @@ export class OssVulnerabilityProvider extends AnalysisTreeNodeProvder {
         vulnerabilityNodes.push(
           new TreeNode({
             text: `${vuln.packageName}@${vuln.version} - ${vuln.title}`,
-            icon: OssVulnerabilityProvider.getSeverityIcon(vuln.severity),
+            icon: OssVulnerabilityTreeProvider.getSeverityIcon(vuln.severity),
             internal: {
-              severity: OssVulnerabilityProvider.getSeverityComparatorIndex(vuln.severity),
+              severity: OssVulnerabilityTreeProvider.getSeverityComparatorIndex(vuln.severity),
             },
             command: {
               command: SNYK_OPEN_ISSUE_COMMAND,
@@ -85,16 +85,16 @@ export class OssVulnerabilityProvider extends AnalysisTreeNodeProvder {
       }
 
       vulnerabilityNodes.sort(this.compareNodes);
-      const fileSeverity = OssVulnerabilityProvider.getFileSeverity(counts);
+      const fileSeverity = OssVulnerabilityTreeProvider.getFileSeverity(counts);
 
       const fileNode = new TreeNode({
         text: fileResult.displayTargetFile,
         description: this.getIssueDescriptionText(fileResult.projectName, fileVulnerabilities),
-        icon: OssVulnerabilityProvider.getSeverityIcon(fileSeverity),
+        icon: OssVulnerabilityTreeProvider.getSeverityIcon(fileSeverity),
         children: vulnerabilityNodes,
         internal: {
           nIssues: vulnerabilityNodes.length,
-          severity: OssVulnerabilityProvider.getSeverityComparatorIndex(fileSeverity),
+          severity: OssVulnerabilityTreeProvider.getSeverityComparatorIndex(fileSeverity),
         },
       });
       nodes.push(fileNode);
