+++ src/snyk/snykOss/views/suggestion/ossSuggestionWebviewScript.ts
